update: skip evals without required jobs (#58)
This also prints which jobs cause an eval to be skipped, splits out the
release job definition, and assumes an eval missing required jobs is the
same as an eval with failing jobs.

Because evals without required jobs are usually due to eval errors, this
should be an okay assumption. If for some reason some of our required
jobs are intentionally no longer in the jobset, the script will
eventually start hard failing and sending notifications (when it can no
longer find adequate evals), and we can fix it at that time.